ref: optimizations
